Refactor: Relocate main activity files and update ViewModel
